backend/requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.2.1
sqlparse==0.5.4
tzdata==2025.2
